pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
mysqlclient = "^2.1.1"
dj-database-url = "^1.2.0"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
celery = "^5.2.7"
redis = "^4.4.0"
drf-yasg = "^1.21.4"
django-rest-passwordreset = "^1.3.0"
django-imagekit = "^4.1.0"
pillow = "^9.4.0"
django-filter = "^22.1"



[tool.poetry.group.dev.dependencies]
factory-boy = "^3.2.1"
pytest = "^7.2.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
